+++ src/external_signer.cpp
@@ -5,8 +5,8 @@
 #include <chainparams.h>
 #include <core_io.h>
 #include <psbt.h>
+#include <util/run_command.h>
 #include <util/strencodings.h>
-#include <util/system.h>
 #include <external_signer.h>
 
 #include <algorithm>
